+++ frontend/src/app/app.module.ts
@@ -80,6 +80,7 @@ import { ChooseSourceComponent } from './projects/models/model-source/choose-sou
 import { ManageGitModelComponent } from './projects/models/model-source/git/manage-git-model/manage-git-model.component';
 import { ManageT4CModelComponent } from './projects/models/model-source/t4c/manage-t4c-model/manage-t4c-model.component';
 import { ModelWrapperComponent } from './projects/models/model-wrapper/model-wrapper.component';
+import { ModelComplexityBadgeComponent } from './projects/project-detail/model-overview/model-complexity-badge/model-complexity-badge.component';
 import { ModelOverviewComponent } from './projects/project-detail/model-overview/model-overview.component';
 import { ProjectDetailsComponent } from './projects/project-detail/project-details.component';
 import { ProjectMetadataComponent } from './projects/project-detail/project-metadata/project-metadata.component';
@@ -158,6 +159,7 @@ import { SettingsComponent } from './settings/settings.component';
     MatCheckboxLoaderComponent,
     MatIconComponent,
     MatListSkeletonLoaderComponent,
+    ModelComplexityBadgeComponent,
     ModelDescriptionComponent,
     ModelDetailComponent,
     ModelDiagramDialogComponent,