+++ frontend/src/app/general/error-handling/error-handling.interceptor.ts
@@ -4,6 +4,7 @@
  */
 
 import {
+  HttpContextToken,
   HttpErrorResponse,
   HttpEvent,
   HttpEventType,
@@ -15,6 +16,10 @@ import { Injectable } from '@angular/core';
 import { getReasonPhrase } from 'http-status-codes';
 import { Observable, tap, map, from, catchError } from 'rxjs';
 import { ToastService } from 'src/app/helpers/toast/toast.service';
+
+// Skips the automated error handling.
+// When this option is set, the error messages from the backend are not auto-printed as toast message
+export const SKIP_ERROR_HANDLING = new HttpContextToken<boolean>(() => false);
 
 @Injectable({
   providedIn: 'root',
@@ -30,6 +35,9 @@ export class ErrorHandlingInterceptor implements HttpInterceptor {
       catchError(this.constructErrorDetailTransformationObservable),
       tap({
         next: (event: HttpEvent<any>) => {
+          if (this.isErrorHandlingSkipped(request)) {
+            return;
+          }
           if (event.type == HttpEventType.Response) {
             const body = event.body;
             if (body?.errors) {
@@ -61,6 +69,10 @@ export class ErrorHandlingInterceptor implements HttpInterceptor {
             return;
           }
 
+          if (this.isErrorHandlingSkipped(request)) {
+            return;
+          }
+
           if (err.error && err.error.detail) {
             let detail = err.error.detail;
             if (Array.isArray(detail)) {
@@ -72,11 +84,11 @@ export class ErrorHandlingInterceptor implements HttpInterceptor {
                 );
               }
             } else if (detail.reason) {
-              if (Array.isArray(detail.reason)) {
-                detail.reason = detail.reason.join(' ');
-              }
               // User defined error
-              this.toastService.showError('An error occurred!', detail.reason);
+              this.toastService.showError(
+                'An error occurred!',
+                ErrorHandlingInterceptor.getErrorReason(detail)
+              );
             }
           } else if (err.status === 0) {
             this.toastService.showError(
@@ -101,6 +113,14 @@ export class ErrorHandlingInterceptor implements HttpInterceptor {
         return event;
       })
     );
+  }
+
+  static getErrorReason(detail: ErrorDetail): string {
+    if (Array.isArray(detail.reason)) {
+      return detail.reason.join(' ');
+    }
+
+    return detail.reason || '';
   }
 
   constructErrorDetailTransformationObservable(err: any): Observable<any> {
@@ -136,4 +156,15 @@ export class ErrorHandlingInterceptor implements HttpInterceptor {
 
     throw err;
   }
+
+  private isErrorHandlingSkipped(request: HttpRequest<any>) {
+    return request.context.get(SKIP_ERROR_HANDLING);
+  }
 }
+
+export type ErrorDetail = {
+  err_code?: string;
+  title?: string;
+  reason?: string | string[];
+  technical?: string;
+};